core-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-jose==3.3.0
bcrypt==4.2.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
httpx==0.27.2
python-multipart==0.0.12
alembic==1.13.3
pycryptodome==3.20.0
